Use is-deeply to compare Sets
Sets are unordered, so stringifying a Set can give any result. `is`
test routine just compares strings. Use `is-deeply` instead to make
sure that the test is successful independent from the stringification
order (which is randomized in newer Rakudos).

Also changes `is` to `is-deeply` for testing array contents (not
strictly necessary in this case, but it is good practice).